pyproject.toml
[tool.poetry]
name = "programming-llm"
version = "0.1.0"
description = ""
authors = ["al002 <al002@disroot.org>"]
readme = "README.md"
packages = [{include = "programming_llm"}]

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.4.25"
langchain = "^0.0.107"
grpcio = "^1.51.3"
grpcio-tools = "^1.51.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
